refactor: tira testes do index.ts
